chore: address code rabbit comments
