+++ libs/application/templates/rental-agreement/src/forms/draft/rentalHousingSubsections/rentalHousingPropertyInfo.ts
@@ -8,13 +8,13 @@ import {
 } from '@island.is/application/core'
 import { SubSection } from '@island.is/application/types'
 import { RentalHousingCategoryClass, applicationAnswers } from '../../../shared'
-import {
-  getPropertyTypeOptions,
-  getPropertyClassOptions,
-  getPropertyClassGroupOptions,
-} from '../../../utils/utils'
 import { Routes, RentalHousingCategoryTypes } from '../../../utils/enums'
 import * as m from '../../../lib/messages'
+import {
+  getPropertyClassGroupOptions,
+  getPropertyClassOptions,
+  getPropertyTypeOptions,
+} from '../../../utils/options'
 
 const messagesInfo = m.registerProperty.info
 const messagesCategory = m.registerProperty.category
